refactoring to separate utils, data, scripts
