+++ noiseandsilence.egg-info/PKG-INFO
@@ -1,0 +1,10 @@
+Metadata-Version: 1.0
+Name: noiseandsilence
+Version: 0.1.0
+Summary: noiseandsilnce research submission.
+Home-page: UNKNOWN
+Author: Gurpreet Johl
+Author-email: gurpreetjohl.com
+License: UNKNOWN
+Description: # noiseandsilence
+Platform: UNKNOWN
