+++ noiseandsilence/data/data_sources.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""
+Sources of info
+"""
+import os
+
+SEPARATOR = ' ~~~ '
+
+UK_URL = 'https://www.officialcharts.com/chart-news/all-the-number-1-singles__7931/'
+
+DATA_DIR = r'noiseandsilence\noiseandsilence\data'
+UK_CHART_DATA_PATH = os.path.join(DATA_DIR, 'uk_chart_data_to_20190103')
+LYRICS_PATH = os.path.join(DATA_DIR, 'lyrics_to_20190103')
+RAW_SPOTIFY_DATA_PATH = os.path.join(DATA_DIR, 'raw_features_to_20190103')
+PARSED_SPOTIFY_DATA_PATH = os.path.join(DATA_DIR, 'parsed_features_to_20190103')
