+++ noiseandsilence/utils/__init__.py
@@ -1,0 +1,6 @@
+from .utils import (try_else_value, get_soup_from_url, 
+                    get_uk_chart_data,
+                    get_lyrics, 
+                    get_spotify_uri_from_title_and_artist,
+                    get_audio_features_from_uri, 
+                    add_audio_features_to_dataframe)
